+++ ci_tools/check_accidental_draft.py
@@ -17,7 +17,7 @@ if __name__ == '__main__':
 
     for pr_id in prs:
         previous_comments, last_comment, last_date = check_previous_comments()
-        if not (last_comment == draft_comment or last_comment == stale_comment):
+        if last_comment not in (draft_comment, stale_comment):
             #TODO: Uncomment before merging
             #leave_comment(pr_id, draft_comment, True)
             pass
